增加C#实体类导出

+++ main.go
@@ -1,5 +1,7 @@
 package main
 
+//mac下构建windows命令：GOOS=windows GOARCH=amd64 go build
+
 import (
 	"bytes"
 	"encoding/json"
@@ -12,7 +14,7 @@ import (
 	"github.com/360EntSecGroup-Skylar/excelize"
 )
 
-var rowNum = map[string]int{"Key": 2, "Type": 1, "DataStart": 4} // Key:字段key所在行 Type:字段类型所在行 DataStart:数据开始的行 (行数从0开始)
+var rowNum = map[string]int{"Key": 0, "TypeTS": 1, "TypeCS": 2, "DataStart": 4} // Key:字段key所在行 Type:字段类型所在行 DataStart:数据开始的行 (行数从0开始)
 
 var excelMap = map[string]string{}
 
@@ -27,44 +29,54 @@ func walkFunc(path string, info os.FileInfo, err error) error {
 }
 
 func main() {
-	entityHeader := ""
-	entityResult := ""
+	entityHeaderTS := ""
+	entityResultTS := ""
+	entityHeaderCS := "using System.Collections;\nusing System.Collections.Generic;\n\n"
+	entityResultCS := ""
 	currentDir, _ := os.Getwd()
-	os.RemoveAll(currentDir + "\\out\\")
+	outPath := currentDir + "/out/"
+	os.RemoveAll(outPath)
 	filepath.Walk(currentDir, walkFunc)
 	for name, path := range excelMap {
 		dataResultMap := map[string]interface{}{}
-		entityHeader += "export interface " + name + "   {\n"
-		dataMap, entityStr := readExcel(path)
-		entityResult += entityStr
+		entityHeaderTS += "export interface " + name + "   {\n"
+		entityHeaderCS += "public class " + name + "   \n{\n"
+		dataMap, entityStrTS, entityStrCS := readExcel(path)
+		entityResultTS += entityStrTS
+		entityResultCS += entityStrCS
 		for key, value := range dataMap {
 			dataResultMap[key] = value
-			entityHeader += ("    " + key + ": { [id: number]: " + key + " };\n")
+			entityHeaderTS += ("    " + key + ": { [id: number]: " + key + " };\n")
+			entityHeaderCS += ("    public Dictionary<string," + key + "> " + key + ";\n")
 		}
 		byteBuf := bytes.NewBuffer([]byte{})
 		encoder := json.NewEncoder(byteBuf)
 		encoder.SetEscapeHTML(false)
 		err := encoder.Encode(dataResultMap)
 		if err == nil {
-			saveData(byteBuf.String(), currentDir+"\\out\\"+name+".json")
+			saveData(byteBuf.String(), outPath+name+".json")
 		} else {
 			fmt.Println(err.Error())
 		}
-		entityHeader += "}\n\n"
+		entityHeaderTS += "}\n\n"
+		entityHeaderCS += "}\n\n"
 	}
-	saveData(entityHeader+entityResult, currentDir+"\\out\\DataEntity.ts")
+	saveData(entityHeaderTS+entityResultTS, outPath+"DataEntity.ts")
+	saveData(entityHeaderCS+entityResultCS, outPath+"DataEntity.cs")
 	fmt.Println("Over")
 }
 
-func readExcel(path string) (map[string]map[string]map[string]interface{}, string) {
+func readExcel(path string) (map[string]map[string]map[string]interface{}, string, string) {
 	file, err := excelize.OpenFile(path)
 	if err == nil {
 		obj := map[string]map[string]map[string]interface{}{}
-		entityStr := ""
+		entityStrTS := ""
+		entityStrCS := ""
 		sheetMap := file.GetSheetMap()
 		for _, sheetName := range sheetMap {
 			keyArr := []string{}
-			typeArr := []string{}
+			typeArrTS := []string{}
+			typeArrCS := []string{}
 			sheetKey := strings.Title(sheetName)
 			obj[sheetKey] = map[string]map[string]interface{}{}
 			subObj := obj[sheetKey]
@@ -74,37 +86,53 @@ func readExcel(path string) (map[string]map[string]map[string]interface{}, strin
 					if i == rowNum["Key"] {
 						keyArr = append(keyArr, cellValue)
 					}
-					if i == rowNum["Type"] {
-						typeArr = append(typeArr, cellValue)
+					if i == rowNum["TypeTS"] {
+						typeArrTS = append(typeArrTS, cellValue)
+					}
+					if i == rowNum["TypeCS"] {
+						typeArrCS = append(typeArrCS, cellValue)
 					}
 					if i >= rowNum["DataStart"] {
 						if j == 0 {
 							subObj[cellValue] = map[string]interface{}{}
 						}
-						if typeArr[j] == "none" || typeArr[j] == "" {
+						if typeArrTS[j] == "none" {
 							continue
 						}
-						subObj[row[0]][keyArr[j]] = getValueByType(cellValue, typeArr[j])
+						if typeArrTS[j] != "" || typeArrCS[j] != "" {
+							subObj[row[0]][keyArr[j]] = getValueByType(cellValue, typeArrTS[j])
+						}
 					}
 				}
 			}
-			entityStr += "export interface " + sheetKey + "  {\n"
-			if len(keyArr) == len(typeArr) {
-				for k, v := range typeArr {
+			entityStrTS += "export interface " + sheetKey + "  {\n"
+			if len(keyArr) == len(typeArrTS) {
+				for k, v := range typeArrTS {
 					if v == "none" || v == "" {
 						continue
 					}
-					entityStr += "    " + keyArr[k] + ": " + typeArr[k] + ";\n"
+					entityStrTS += "    " + keyArr[k] + ": " + typeArrTS[k] + ";\n"
 
 				}
 			}
-			entityStr += "}\n\n"
+			entityStrTS += "}\n\n"
+			entityStrCS += "public class " + sheetKey + "  \n{\n"
+			if len(keyArr) == len(typeArrCS) {
+				for k, v := range typeArrCS {
+					if v == "none" || v == "" {
+						continue
+					}
+					entityStrCS += "    public " + typeArrCS[k] + " " + keyArr[k] + ";\n"
+
+				}
+			}
+			entityStrCS += "}\n\n"
 		}
-		return obj, entityStr
+		return obj, entityStrTS, entityStrCS
 	} else {
 		fmt.Println(err.Error())
 	}
-	return nil, ""
+	return nil, "", ""
 }
 
 func getValueByType(str string, typeStr string) interface{} {
@@ -157,18 +185,26 @@ func checkFile(filePath string) {
 	dir := filepath.Dir(filePath)
 	_, err1 := os.Stat(dir)
 	if err1 != nil {
-		e := os.Mkdir(dir, os.ModeDir)
-		if e != nil {
-			fmt.Println(e.Error())
+		if os.IsNotExist(err1) {
+			e := os.MkdirAll(dir, os.ModeDir)
+			if e != nil {
+				fmt.Println(e.Error())
+			} else {
+				os.Chmod(dir, os.ModePerm)
+			}
+		} else {
+			fmt.Println(err1.Error())
 		}
-	} else {
-
 	}
 	_, err2 := os.Stat(filePath)
 	if err2 != nil {
-		_, e := os.Create(filePath)
-		if e != nil {
-			fmt.Println(e.Error())
+		if os.IsNotExist(err2) {
+			_, e := os.Create(filePath)
+			if e != nil {
+				fmt.Println(e.Error())
+			}
+		} else {
+			fmt.Println(err2.Error())
 		}
 	}
 }
